TCLab
The temperature control lab (TCLab) is an application of feedback control with an Arduino, an LED, two heaters, and two temperature sensors. The heater power output is adjusted to maintain a desired temperature setpoint. Thermal energy from the heater is transferred by conduction, convection, and radiation to the temperature sensor. Heat is also transferred away from the device to the surroundings.

This lab is a resource for model identification and controller development. It is a pocket-sized lab with software in Python, MATLAB, and Simulink for the purpose of reinforcing control theory for students. 

https://apmonitor.com/pdc/index.php/Main/ArduinoTemperatureControl